moved template pagination logic to paginable namespace
