--- app/views/org_admin/templates/_templates_list.html.erb
@@ -1,88 +1,0 @@
-<% # locals: templates, current_org %>
-
-<div class="table-responsive">
-  <table class="table table-hover table-bordered tablesorter">
-    <thead>
-      <% if scope.length > TABLE_FILTER_MIN_ROWS %>
-        <tr>
-          <th colspan="5" class="sorter-false">
-            <%= render(partial: "shared/table_filter", locals: { placeholder: _('Search')}) %>
-          </th>
-        </tr>
-      <% end %>
-      <% if scopes.present? %>
-        <tr><th colspan="5" class="sorter-false table-scope"><ul class="nav navbar-nav">
-          <% scopes.keys.each do |k| %>
-            <% unless k == :dmptemplate_ids %>
-              <li><%= link_to "#{k.capitalize} (#{scopes[k]})", "#{hide_actions ? all_org_admin_templates_path(1) : orgs_org_admin_templates_path(1)}?scope=#{k}", class: 'template-scope', 'data-remote': true %></li>
-            <% end %>
-          <% end %>
-        </ul></th></tr>
-      <% end %>
-      <tr>
-        <th><%= _('Template Name') %></th>
-        <th><%= current_user.can_super_admin? ? _('Organisation') : _('Description') %></th>
-        <th><%= _('Status') %></th>
-        <th><%= _('Edited Date') %></th>
-        <% if !hide_actions %><th class="sorter-false"></th><% end %>
-      </tr>
-    </thead>
-    <tbody>
-    <% scope.each do |template| %>
-      <tr>
-        <td>
-          <%= template.title %>
-        </td>
-          <td>
-            <%= current_user.can_super_admin? ? template.org.name : raw(template.description) %>
-          </td>
-          <td>
-            <% if !published[template.dmptemplate_id].present? %>
-              <%= _('Unpublished') %>
-            <% elsif template.dirty? %>
-              <%= _('Unpublished changes') %>
-            <% else %>
-              <%= _('Published') %>
-            <% end %>
-          </td>
-          <td>
-            <% last_temp_updated = template.updated_at %>
-            <%= l last_temp_updated.to_date, formats: :short %>
-          </td>
-          <% if !hide_actions %>
-            <td>
-              <div class="dropdown<%= current_org == template.org.id ? '' : ' hide'%>">
-                <% # Span used to trigger the display of the action dropdowns via JS %>
-                <span class="super-admin-template-org hide"><%= template.org.id %></span>
-
-                <button class="btn btn-link dropdown-toggle" type="button"
-                        data-toggle="dropdown"
-                        aria-haspopup="true" aria-expanded="true">
-                  <%= _('Actions') %><span class="caret"></span>
-                </button>
-                <ul class="dropdown-menu">
-                  <li><%= link_to _('Edit'), edit_org_admin_template_path(id: template.id, edit: "true") %></li>
-                  <li><%= link_to _('History'), history_org_admin_template_path(id: template.id) %></li>
-                  <% if !published[template.dmptemplate_id].present? %>
-                    <li><%= link_to _('Publish'), publish_org_admin_template_path(template) %></li>
-                  <% elsif template.dirty? %>
-                    <li><%= link_to _('Publish changes'), publish_org_admin_template_path(template) %></li>
-                  <% else %>
-                    <li><%= link_to _('Unpublish'), unpublish_org_admin_template_path(template) %></li>
-                  <% end %>
-                  <li><%= link_to _('Copy'), copy_org_admin_template_path(template) %></li>
-                  <% if template.plans.length <= 0 %>
-                    <li>
-                      <%= link_to _('Remove'), org_admin_template_path(id: template.id), 'data-method': 'delete', rel: 'nofollow',
-                                  'data-confirm': _('Are you sure you want to remove "%{template_title}"?') % { template_title: template.title} %>
-                    </li>
-                  <% end %>
-                </ul>
-              </div>
-           </td>
-         <% end %>
-       </tr>
-      <% end %>
-    </tbody>
-  </table>
-</div>
