+++ app/views/org_admin/templates/index.html.erb
@@ -44,8 +44,8 @@
         <div id="all-templates" role="tabpanel" class="tab-pane active">
           <h2><%= _('All Templates') %></h2>
           <%= paginable_renderise(
-            partial: 'templates_list',
-            controller: 'org_admin/templates',
+            partial: 'paginable/templates/all',
+            controller: 'paginable/templates',
             action: 'all', 
             scope: all_templates,
             locals: {current_org: current_org.id, published: published, scopes: scopes[:all], hide_actions: true}) %>
@@ -54,8 +54,8 @@
       <div id="organisation-templates" role="tabpanel" class="tab-pane<%= !current_user.can_super_admin? ? ' active' : '' %>">
         <h2><%= current_user.can_super_admin? ? _('%{org_name} Templates') % { org_name: current_user.org.name } : _('Own Templates') %></h2>
         <%= paginable_renderise(
-          partial: 'templates_list',
-          controller: 'org_admin/templates',
+          partial: 'paginable/templates/orgs',
+          controller: 'paginable/templates',
           action: 'orgs', 
           scope: own_templates,
           locals: {current_org: current_org.id, published: published, scopes: scopes[:orgs], hide_actions: false}) %>
@@ -65,8 +65,8 @@
         <div id="funder-templates" role="tabpanel" class="tab-pane">
           <h2><%= _('Customizable Templates') %></h2>
           <%= paginable_renderise(
-            partial: 'funder_templates_list',
-            controller: 'org_admin/templates',
+            partial: 'paginable/templates/funders',
+            controller: 'paginable/templates',
             action: 'funders', 
             scope: customizable_templates,
             locals: {current_org: current_org.id, customizations: customized_templates, published: published, scopes: scopes[:funders]}) %>
