+++ app/views/paginable/templates/_all.html.erb
@@ -1,0 +1,50 @@
+<% # locals: templates, current_org %>
+
+<div class="table-responsive">
+  <table class="table table-hover table-bordered tablesorter">
+    <thead>
+      <% if scopes.present? %>
+        <tr><th colspan="4" class="sorter-false table-scope"><ul class="nav navbar-nav">
+          <% scopes.keys.each do |k| %>
+            <% unless k == :dmptemplate_ids %>
+              <li><%= link_to "#{k.capitalize} (#{scopes[k]})", 
+                              "#{all_paginable_templates_path(1)}?scope=#{k}", 
+                              class: 'template-scope', 'data-remote': true %></li>
+            <% end %>
+          <% end %>
+        </ul></th></tr>
+      <% end %>
+      <tr>
+        <th><%= _('Template Name') %></th>
+        <th><%= _('Organisation') %></th>
+        <th><%= _('Status') %></th>
+        <th><%= _('Edited Date') %></th>
+      </tr>
+    </thead>
+    <tbody>
+    <% scope.each do |template| %>
+      <tr>
+        <td>
+          <%= template.title %>
+        </td>
+          <td>
+            <%= template.org.name %>
+          </td>
+          <td>
+            <% if !published[template.dmptemplate_id].present? %>
+              <%= _('Unpublished') %>
+            <% elsif template.dirty? %>
+              <%= _('Unpublished changes') %>
+            <% else %>
+              <%= _('Published') %>
+            <% end %>
+          </td>
+          <td>
+            <% last_temp_updated = template.updated_at %>
+            <%= l last_temp_updated.to_date, formats: :short %>
+          </td>
+       </tr>
+      <% end %>
+    </tbody>
+  </table>
+</div>
