+++ app/views/paginable/templates/_funders.html.erb
@@ -3,19 +3,12 @@
 <div class="table-responsive">
   <table class="table table-hover table-bordered tablesorter">
     <thead>
-      <% if scope.length > TABLE_FILTER_MIN_ROWS %>
-        <tr>
-          <th colspan="5" class="sorter-false">
-            <%= render(partial: "shared/table_filter", locals: { placeholder: _('Search')}) %>
-          </th>
-        </tr>
-      <% end %>
       <% if scopes.present? %>
         <tr><th colspan="5" class="sorter-false table-scope"><ul class="nav navbar-nav">
           <% scopes.keys.each do |k| %>
             <% unless k == :dmptemplate_ids %>
               <% label = (['published', 'unpublished'].include?(k.to_s) ? _('Customization %{publication_status}') % { publication_status: k } : k.capitalize) %>
-              <li><%= link_to "#{label} (#{scopes[k]})", "#{funders_org_admin_templates_path(1)}?scope=#{k}", class: 'template-scope', 'data-remote': true %></li>
+              <li><%= link_to "#{label} (#{scopes[k]})", "#{funders_paginable_templates_path(1)}?scope=#{k}", class: 'template-scope', 'data-remote': true %></li>
             <% end %>
           <% end %>
         </ul></th></tr>
